requirements.txt
# Core framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pyyaml

# AI / ML
openai
onnxruntime
langchain

# Automation & capture
pyautogui
pynput
pillow
opencv-python
pytesseract
easyocr

# Scheduling & tasks
schedule
celery>=5.2.0
redis>=4.0.0
watchdog
sse_starlette
croniter>=1.3.0

# Database & storage
psycopg2-binary>=2.9.0
pymongo
boto3

# Integrations
twilio
slack-sdk

# Authentication / Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# Utilities
python-multipart
requests>=2.31.0

# Added from the code block
sqlmodel

# New dependencies for enhanced functionality
websockets>=10.0
aiofiles>=0.8.0

# Automation extras (low-level input capture for recording)
evdev-binary

# ------------------------------------------------------------------- 
# Retrieval-Augmented Generation (RAG) Support
# ------------------------------------------------------------------- 
# Vector stores
faiss-cpu>=1.7.4            # Local similarity-search index
chromadb>=0.4.24            # Lightweight, serverless vector DB (optional)

# Text splitting / token utilities
tiktoken>=0.5.1

# Document loaders & parsers
unstructured>=0.10.18       # Robust loader for many doc types (PDF, HTML, etc.)
pypdf>=3.16.4               # PDF parsing (used by PyPDFLoader)

# Embeddings back-ends (OpenAI already listed above)
sentence-transformers>=2.2.2   # Local embedding models (optional)

# Production server
gunicorn>=20.1.0

# Testing dependencies
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-cov>=4.0.0
pytest-xdist>=3.3.0
mutmut>=2.4.3
responses>=0.25.0

# Browser automation & playwright helpers
playwright>=1.42.0
pytest-playwright>=0.4.4

# Security dependencies
restrictedpython>=5.2

# Load testing dependencies
locust>=2.17.0

# Monitoring dependencies
schedule>=1.2.0

# Redis client
redis>=5.0.0
